+++ utils/versionCore.ts
