+++ utils/versionSolana.ts
@@ -6,7 +6,7 @@ import {
   getVersionFromStdout,
   ResolvedVersion,
   Version,
-} from './version-core';
+} from './versionCore';
 
 export async function detectSolanaVersion(
   language: Language
